Refactors automatic bookmark generation and clipboard copy logic into a shared helper function. This new helper is now used by the signature settings and system link dialogs.
- feat(mapper): Add handleAutoBookmark helper for bookmark generation and copying
- refactor(mapper): Use handleAutoBookmark helper in SystemLinkSignatureDialog
- refactor(mapper): Use handleAutoBookmark helper in SignatureSettings
